Refactor FCesiumTextureResource into two classes.
