+++ Source/CesiumRuntime/Private/CesiumEncodedMetadataConversions.cpp
@@ -133,7 +133,7 @@ template <typename T>
 void coerceAndEncodeArrays(
     const FCesiumPropertyTablePropertyDescription& propertyDescription,
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData,
+    const gsl::span<std::byte>& textureData,
     size_t pixelSize) {
   int64 propertySize =
       UCesiumPropertyTablePropertyBlueprintLibrary::GetPropertySize(property);
@@ -178,7 +178,7 @@ void coerceAndEncodeArrays(
 template <typename T>
 void coerceAndEncodeScalars(
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData) {
+    const gsl::span<std::byte>& textureData) {
   int64 propertySize =
       UCesiumPropertyTablePropertyBlueprintLibrary::GetPropertySize(property);
   if (textureData.size() < propertySize * sizeof(T)) {
@@ -206,7 +206,7 @@ void coerceAndEncodeScalars(
 template <typename T>
 void coerceAndEncodeVec2s(
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData,
+    const gsl::span<std::byte>& textureData,
     size_t pixelSize) {
   int64 propertySize =
       UCesiumPropertyTablePropertyBlueprintLibrary::GetPropertySize(property);
@@ -250,7 +250,7 @@ void coerceAndEncodeVec2s(
 template <typename T>
 void coerceAndEncodeVec3s(
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData,
+    const gsl::span<std::byte>& textureData,
     size_t pixelSize) {
   int64 propertySize =
       UCesiumPropertyTablePropertyBlueprintLibrary::GetPropertySize(property);
@@ -293,7 +293,7 @@ void coerceAndEncodeVec3s(
 template <typename T>
 void coerceAndEncodeVec4s(
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData,
+    const gsl::span<std::byte>& textureData,
     size_t pixelSize) {
   int64 propertySize =
       UCesiumPropertyTablePropertyBlueprintLibrary::GetPropertySize(property);
@@ -376,7 +376,7 @@ bool CesiumEncodedMetadataCoerce::canEncode(
 void CesiumEncodedMetadataCoerce::encode(
     const FCesiumPropertyTablePropertyDescription& propertyDescription,
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData,
+    const gsl::span<std::byte>& textureData,
     size_t pixelSize) {
   if (propertyDescription.PropertyDetails.bIsArray) {
     if (propertyDescription.EncodingDetails.ComponentType ==
@@ -489,7 +489,7 @@ glm::u8vec3 getRgbColorFromString(const FString& rgbString) {
 template <typename T>
 void parseAndEncodeColors(
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData,
+    const gsl::span<std::byte>& textureData,
     size_t pixelSize) {
   int64 propertySize =
       UCesiumPropertyTablePropertyBlueprintLibrary::GetPropertySize(property);
@@ -546,7 +546,7 @@ bool CesiumEncodedMetadataParseColorFromString::canEncode(
 void CesiumEncodedMetadataParseColorFromString::encode(
     const FCesiumPropertyTablePropertyDescription& propertyDescription,
     const FCesiumPropertyTableProperty& property,
-    gsl::span<std::byte>& textureData,
+    const gsl::span<std::byte>& textureData,
     size_t pixelSize) {
   if (propertyDescription.EncodingDetails.ComponentType ==
       ECesiumEncodedMetadataComponentType::Uint8) {
